Fix circular import issue and get tests running
